Cambios en la función set_currency
