+++ lib/siwapp/invoices.ex
@@ -185,14 +185,4 @@ defmodule Siwapp.Invoices do
   def change_item(%Item{} = item, attrs \\ %{}) do
     Item.changeset(item, attrs)
   end
-
-  @spec set_currency(float | integer, atom | binary) :: binary
-  def set_currency(value, nil), do: set_currency(value, :USD)
-
-  def set_currency(value, currency) do
-    value
-    |> round()
-    |> Money.new(currency)
-    |> Money.to_string()
-  end
 end
