+++ lib/siwapp_web/controllers/iframe_controller.ex
@@ -19,9 +19,7 @@ defmodule SiwappWeb.IframeController do
       invoice_eval_data ++
         [
           have_discount?: have_items_discount?(invoice.items),
-          status: Invoices.status(invoice),
-          gross_amount_curr: Invoices.set_currency(invoice.gross_amount, invoice.currency),
-          net_amount_curr: Invoices.set_currency(invoice.net_amount, invoice.currency)
+          status: Invoices.status(invoice)
         ]
 
     str_template = EEx.eval_string(template, all_eval_data)
