+++ lib/siwapp_web/live/invoices_live/index.html.heex
@@ -57,14 +57,14 @@
 
           <% :pending -> %>
             <td class="table-span"><span class="tag is-warning">Pending</span></td>
-            <td><%= Invoices.set_currency(invoice.gross_amount, invoice.currency) %></td>
+            <td><%= PageView.set_currency(invoice.gross_amount, invoice.currency) %></td>
 
           <% :past_due -> %>
             <td class="table-span"><span class="tag is-danger">Past_Due</span></td>
-            <td><%= Invoices.set_currency(invoice.gross_amount, invoice.currency) %></td>
+            <td><%= PageView.set_currency(invoice.gross_amount, invoice.currency) %></td>
         <% end %>
 
-        <td><%= Invoices.set_currency(invoice.gross_amount, invoice.currency) %></td>
+        <td><%= PageView.set_currency(invoice.gross_amount, invoice.currency) %></td>
       </tr>
     <% end %>
   </tbody>
