+++ lib/siwapp_web/views/iframe_view.ex
@@ -1,0 +1,7 @@
+defmodule SiwappWeb.IframeView do
+  use SiwappWeb, :view
+
+  alias SiwappWeb.PageView
+
+  def set_currency(value, currency), do: PageView.set_currency(value, currency)
+end
